+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/BaseSpawnerMixin.java
@@ -8,6 +8,8 @@ import net.minecraft.world.level.BaseSpawner;
 import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(BaseSpawner.class)
 public abstract class BaseSpawnerMixin {
@@ -27,8 +29,13 @@ public abstract class BaseSpawnerMixin {
 			at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/BaseSpawner;isNearPlayer(Lnet/minecraft/world/level/Level;Lnet/minecraft/core/BlockPos;)Z")
 	)
 	private boolean clientTick(boolean isPlayerNearby, Level level, BlockPos blockPos) {
-		if (isPlayerNearby && Spawners.onSpawnerClientTick((BaseSpawner) (Object) this))
+		if (isPlayerNearby && Spawners.onSpawnerClientTick((BaseSpawner) (Object) this, level))
 			return false;
 		return isPlayerNearby;
+	}
+
+	@Inject(method = "delay", at = @At("RETURN"))
+	private void onDelay(Level level, BlockPos pos, CallbackInfo ci) {
+		Spawners.onSpawnerDelaySet((BaseSpawner) (Object) this, level, pos);
 	}
 }