+++ src/main/java/insane96mcp/iguanatweaksreborn/module/world/spawners/SpawnerStatusSync.java
@@ -1,4 +1,4 @@
-package insane96mcp.iguanatweaksreborn.network.message;
+package insane96mcp.iguanatweaksreborn.module.world.spawners;
 
 import insane96mcp.iguanatweaksreborn.module.world.spawners.capability.SpawnerData;
 import insane96mcp.iguanatweaksreborn.module.world.spawners.capability.SpawnerDataImpl;
@@ -13,19 +13,25 @@ import java.util.function.Supplier;
 public class SpawnerStatusSync {
 	BlockPos pos;
 	SpawnerDataImpl spawnerData;
+	int delay;
+	int requiredPlayerRange;
 
-	public SpawnerStatusSync(BlockPos pos, SpawnerDataImpl spawnerData) {
+	public SpawnerStatusSync(BlockPos pos, SpawnerDataImpl spawnerData, int delay, int requiredPlayerRange) {
 		this.pos = pos;
 		this.spawnerData = spawnerData;
+		this.delay = delay;
+		this.requiredPlayerRange = requiredPlayerRange;
 	}
 
 	public static void encode(SpawnerStatusSync pkt, FriendlyByteBuf buf) {
 		buf.writeBlockPos(pkt.pos);
 		pkt.spawnerData.toNetwork(buf);
+		buf.writeInt(pkt.delay);
+		buf.writeInt(pkt.requiredPlayerRange);
 	}
 
 	public static SpawnerStatusSync decode(FriendlyByteBuf buf) {
-		return new SpawnerStatusSync(buf.readBlockPos(), SpawnerDataImpl.fromNetwork(buf));
+		return new SpawnerStatusSync(buf.readBlockPos(), SpawnerDataImpl.fromNetwork(buf), buf.readInt(), buf.readInt());
 	}
 
 	public static void handle(final SpawnerStatusSync message, Supplier<NetworkEvent.Context> ctx) {
@@ -36,6 +42,8 @@ public class SpawnerStatusSync {
 					iSpawner.setDisabled(message.spawnerData.isDisabled());
 					iSpawner.setEmpowered(message.spawnerData.isEmpowered());
 				});
+				spawnerBlockEntity.getSpawner().spawnDelay = message.delay;
+				spawnerBlockEntity.getSpawner().requiredPlayerRange = message.delay;
 			}
 		});
 		ctx.get().setPacketHandled(true);