+++ changelog.md
@@ -2,8 +2,9 @@
 
 ## Upcoming
 * Reworked spawners
-  * Spawners now generate empowered and will spawn mobs really fast  
+  * Spawners now generate empowered and will spawn mobs faster  
     After 24 mobs spawned the spawner will slowdown the spawning but still remain active and award experience and loot
+  * Also increased the range for activation from 16 to 32
   * Spawners no longer disable themselves after various mobs spawned and no longer give bonus experience when broken
 * No Hunger can now be enabled/disable with a gamerule
 * Tools no longer take 2 damage when hurting entities